requirements.txt
websockets
pyserial
pygame